chore: rename project to sparv-testing
