+++ .github/workflows/release.yml
@@ -109,7 +109,7 @@ jobs:
       - run: rm -r src
       - run: pip install typing-extensions
       - run: pip install -r tests/requirements-testing.lock
-      - run: pip install sparv-pipeline-testing --no-index --no-deps --find-links dist --force-reinstall
+      - run: pip install sparv-testing --no-index --no-deps --find-links dist --force-reinstall
       - run: pytest
 
   # https://github.com/marketplace/actions/alls-green#why used for branch protection checks
