+++ Makefile
@@ -57,8 +57,8 @@ help:
 	@echo ""
 
 PLATFORM := `uname -o`
-REPO := sparv-pipeline-testing
-PROJECT_SRC := src/sparv_pipeline_testing
+REPO := sparv-testing
+PROJECT_SRC := src/sparv_testing
 
 ifeq (${VIRTUAL_ENV},)
   VENV_NAME = .venv
