+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "sparv-pipeline-testing"
+name = "sparv-testing"
 version = "0.1.7"
 description = "Testing utilities for sparv-pipeline."
 readme = "README.md"
@@ -8,10 +8,8 @@ authors = [
     { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
 ]
 license = { text = "MIT" }
-requires-python = ">=3.9"
-dependencies = [
-    "sparv-pipeline>=5.2.0",
-]
+requires-python = ">=3.11"
+dependencies = ["sparv>=5.3.1"]
 
 [build-system]
 requires = ["hatchling"]
