--- src/sparv_pipeline_testing/__init__.py
@@ -1,6 +1,0 @@
-"""Testing utilities for sparv-pipeline."""
-
-from sparv_pipeline_testing.annotations import MockAnnotation
-from sparv_pipeline_testing.outputs import MemoryOutput
-
-__all__ = ["MemoryOutput", "MockAnnotation"]
