+++ src/sparv_testing/__init__.py
@@ -1,0 +1,6 @@
+"""Testing utilities for sparv-pipeline."""
+
+from sparv_testing.annotations import MockAnnotation
+from sparv_testing.outputs import MemoryOutput
+
+__all__ = ["MemoryOutput", "MockAnnotation"]
