+++ tests/requirements-testing.lock
@@ -3,35 +3,29 @@
 appdirs==1.4.4
     # via
     #   snakemake
-    #   sparv-pipeline
-argcomplete==3.1.6
-    # via sparv-pipeline
+    #   sparv
+argcomplete==3.6.2
+    # via sparv
 attrs==23.2.0
     # via
     #   jsonschema
     #   referencing
 certifi==2024.2.2
     # via requests
-cffi==1.16.0
-    # via cryptography
 charset-normalizer==3.3.2
-    # via
-    #   pdfminer-six
-    #   requests
+    # via requests
 click==8.1.7
     # via nltk
 configargparse==1.7
     # via snakemake
 connection-pool==0.0.3
     # via snakemake
-cryptography==42.0.7
-    # via pdfminer-six
 datrie==0.8.2
     # via snakemake
 docutils==0.20.1
     # via snakemake
 docx2python==1.27.1
-    # via sparv-pipeline
+    # via sparv
 dpath==2.1.6
     # via yte
 emoji==2.12.1
@@ -52,8 +46,6 @@ humanfriendly==10.0
     # via snakemake
 idna==3.7
     # via requests
-importlib-metadata==6.8.0
-    # via sparv-pipeline
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.4
@@ -62,11 +54,11 @@ jinja2==3.1.4
     #   torch
 joblib==1.4.2
     # via nltk
-jsonschema==4.20.0
+jsonschema==4.23.0
     # via
     #   nbformat
     #   snakemake
-    #   sparv-pipeline
+    #   sparv
 jsonschema-specifications==2023.12.1
     # via jsonschema
 jupyter-core==5.7.2
@@ -84,9 +76,11 @@ mypy-extensions==1.0.0
 nbformat==5.10.4
     # via snakemake
 networkx==3.1
-    # via torch
-nltk==3.8.1
-    # via sparv-pipeline
+    # via
+    #   stanza
+    #   torch
+nltk==3.9.1
+    # via sparv
 numpy==1.24.4
     # via stanza
 nvidia-cublas-cu12==12.1.3.1
@@ -124,13 +118,7 @@ packaging==24.0
     # via
     #   pytest
     #   snakemake
-    #   sparv-pipeline
-pdfminer-six==20221105
-    # via pdfplumber
-pdfplumber==0.10.3
-    # via sparv-pipeline
-pillow==10.3.0
-    # via pdfplumber
+    #   sparv
 plac==1.4.3
     # via yte
 platformdirs==4.2.2
@@ -141,33 +129,31 @@ prompt-toolkit==3.0.45
     # via questionary
 protobuf==3.20.3
     # via
-    #   sparv-pipeline
+    #   sparv
     #   stanza
 psutil==5.9.8
     # via snakemake
 pulp==2.8.0
     # via snakemake
-pycountry==22.3.5
-    # via sparv-pipeline
-pycparser==2.22
-    # via cffi
+pycountry==24.6.1
+    # via sparv
 pygments==2.18.0
     # via rich
 pypdfium2==4.30.0
-    # via pdfplumber
+    # via sparv
 pytest==8.2.1
     # via -r tests/requirements-testing.in
-python-dateutil==2.8.2
-    # via sparv-pipeline
-python-json-logger==2.0.7
-    # via sparv-pipeline
-pyyaml==6.0.1
+python-dateutil==2.9.0
+    # via sparv
+python-json-logger==3.3.0
+    # via sparv
+pyyaml==6.0.2
     # via
     #   snakemake
-    #   sparv-pipeline
+    #   sparv
     #   yte
 questionary==1.10.0
-    # via sparv-pipeline
+    # via sparv
 referencing==0.35.1
     # via
     #   jsonschema
@@ -177,17 +163,20 @@ regex==2024.5.15
 requests==2.32.3
     # via
     #   snakemake
+    #   sparv
     #   stanza
 reretry==0.11.8
     # via snakemake
-rich==13.7.0
-    # via sparv-pipeline
+rich==14.0.0
+    # via
+    #   rich-argparse
+    #   sparv
+rich-argparse==1.7.0
+    # via sparv
 rpds-py==0.18.1
     # via
     #   jsonschema
     #   referencing
-setuptools==70.0.0
-    # via pycountry
 six==1.16.0
     # via python-dateutil
 smart-open==7.0.4
@@ -195,11 +184,11 @@ smart-open==7.0.4
 smmap==5.0.1
     # via gitdb
 snakemake==7.32.3
-    # via sparv-pipeline
-sparv-pipeline==5.2.0
-    # via sparv-pipeline-testing (pyproject.toml)
-stanza==1.5.1
-    # via sparv-pipeline
+    # via sparv
+sparv==5.3.1
+    # via sparv-testing (pyproject.toml)
+stanza==1.10.1
+    # via sparv
 stopit==1.1.2
     # via snakemake
 sympy==1.12.1
@@ -212,7 +201,7 @@ toposort==1.10
     # via snakemake
 torch==2.3.0
     # via
-    #   sparv-pipeline
+    #   sparv
     #   stanza
 tqdm==4.66.4
     # via
@@ -230,7 +219,7 @@ typing-extensions==4.12.0
     #   torch
     #   typing-inspect
 typing-inspect==0.9.0
-    # via sparv-pipeline
+    # via sparv
 urllib3==2.2.1
     # via requests
 wcwidth==0.2.13
@@ -241,5 +230,3 @@ wrapt==1.16.0
     #   snakemake
 yte==1.5.4
     # via snakemake
-zipp==3.19.0
-    # via importlib-metadata
